+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Update local timeline page to work with new API.
+- Implement error handling for "add alias" operation.
 
 ## [1.24.1] - 2023-05-15